+++ frontend/src/Pages/Homepage.js
@@ -1,0 +1,13 @@
+import PropertyCard from "../components/PropertyCard/PropertyCard";
+import { Container } from "@mui/material";
+function Homepage() {
+	return (
+		<Container>
+			<h3>Home Page</h3>
+
+			<PropertyCard />
+		</Container>
+	);
+}
+
+export default Homepage;
