--- frontend/src/components/Breadcrumbs.js
@@ -1,8 +1,0 @@
-
-
-function Breadcrumbs(){
-return(
-    <div>placeholder</div>
-)
-}
-export default Breadcrumbs()
