--- frontend/src/components/Homepage.js
@@ -1,15 +1,0 @@
-import PropertyCard from "./PropertyCard";
-import { Container } from "@mui/material";
-function Homepage(){
-    
-    return(
-        <Container>
-            <h3>Home Page</h3>
-
-            <PropertyCard/>
-        </Container>
-    )
-
-}
-
-export default Homepage;
