--- frontend/src/components/MapComponent.js
@@ -1,5 +1,0 @@
-function MapComponent() {
-	return <div>MAP GOES HERE</div>;
-}
-
-export default MapComponent;
