--- frontend/src/components/map.css
@@ -1,6 +1,0 @@
-
-
-.map-container{
-    height: 50vh;
-    
-}
