--- frontend/src/components/property_data_sample.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9abb966a87cea9891e4433e80d09952a36f24f9e97289e3b1ccc34053dd6a548
-size 3301
